Created Audio Controller
- created an Audio Controller Prefab that transitions between songs based on the state of the game

- put stuff into their proper folders
