+++ Assets/Fonts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 75f770ed60bd0c14fbdd80a468886199
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
