+++ Assets/Scenes/LevelScene.unity
@@ -2237,3 +2237,64 @@ Transform:
   m_Father: {fileID: 1315378140}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!1001 &417824372154392089
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_RootOrder
+      value: 8
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879336, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 417824372965879337, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: _gameStateController
+      value: 
+      objectReference: {fileID: 1070601697}
+    - target: {fileID: 417824372965879350, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
+      propertyPath: m_Name
+      value: Audio Controller
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 79ecfe1f60463bf40bc9410be9bb213f, type: 3}
