+++ Assets/Scripts/TakeDamage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d94560700813a834e9c0c8ac8e086ecc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
