--- Assets/SetPlayerHealthOnAwake.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 43b7df4d74173d44f8da8c687d6a0852
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
