--- Assets/SetSliderColorOnStart.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2d03fec79a50398488374636cee54188
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
