--- Assets/WaveCooldownSliders.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 84d2af308c8c449409c9e54b23ffaaba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
