--- Assets/Wizardry Night SDF.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f596e19679ede1143af416792c39d46f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
